requirements.txt
flask==3.1.2
pillow==11.3.0
torch==2.8.0
torchvision==0.23.0
accelerate==1.10.0
diffusers @ git+https://github.com/huggingface/diffusers@673d4357ff9f085f6f2cd9eebaff23fd1fd9990a
transformers==4.55.4
tokenizers==0.21.4
safetensors==0.6.2
tqdm==4.67.1
huggingface-hub==0.34.4
filelock==3.19.1
fsspec==2025.7.0
requests==2.32.5
pyyaml==6.0.2
regex==2025.7.34
werkzeug==3.1.3
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
blinker==1.9.0
dfloat11==0.5.0